Split session negotiation into focused modules
